Fix WebTransport not connecting after using a proxy
